# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Stellar-Save Backend API
  description: |
    REST API for the Stellar-Save decentralised rotational savings platform.

    ## Authentication
    Most endpoints are public. Endpoints under `/api/user/*` require a Bearer JWT
    obtained from `POST /api/auth/verify`. Pass it as:
    ```
    Authorization: Bearer <token>
    ```

    ## Versioning
    The API is versioned via URL path prefix:
    - `/api/v1/` — stable, production-ready
    - `/api/v2/` — current development version (some routes return 501)

    Unversioned legacy paths (`/health`, `/recommendations`, etc.) are deprecated
    and will be removed on 2027-01-01.

    ## Rate Limiting
    - General endpoints: 100 req / 15 min per IP
    - Auth & admin endpoints: 10 req / 15 min per IP
    - Analytics write endpoints: stricter limits apply (see individual routes)
  version: 1.0.0
  contact:
    name: Stellar-Save
    url: https://github.com/Xoulomon/Stellar-Save
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.stellar-save.app
    description: Production
...
paths:
...
  /api/webhooks:
    post:
      tags: [Webhooks]
      summary: Register a new webhook
      operationId: webhookCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookInput'
      responses:
        '201':
          description: Webhook created (includes generated secret)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      tags: [Webhooks]
      summary: List webhooks for a user
      operationId: webhookList
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Webhooks list (secrets masked)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WebhookSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/webhooks/{id}:
    get:
      tags: [Webhooks]
      summary: Get a specific webhook
      operationId: webhookGet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Webhook details (secret masked)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookSummary'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Webhooks]
      summary: Update a webhook
      operationId: webhookUpdate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookUpdate'
      responses:
        '200':
          description: Webhook updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Webhooks]
      summary: Delete a webhook
      operationId: webhookDelete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Webhook deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Cache ─────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by POST /api/auth/verify
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    WebhookInput:
      type: object
      required: [userId, url, events]
      properties:
        userId:
          type: string
        groupId:
          type: string
          description: Optional — scope webhook to a specific group
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
          example: [GroupCreated, ContributionMade, PayoutExecuted]
        secret:
          type: string
          description: HMAC-SHA256 signing secret (auto-generated if omitted)
        description:
          type: string
    Webhook:
      allOf:
        - $ref: '#/components/schemas/WebhookInput'
        - type: object
          properties:
            id:
              type: string
            isActive:
              type: boolean
            createdAt:
              type: string
              format: date-time
    WebhookSummary:
      allOf:
        - $ref: '#/components/schemas/Webhook'
      description: Same as Webhook but with `secret` omitted for security
    WebhookUpdate:
      type: object
      required: [userId]
      properties:
        userId:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        description:
          type: string